Extract remove action
